# [drci] Optimize sql query for recent workflow jobs