--- a/torchci/rockset/commons/__sql/recent_pr_workflows_query.sql
+++ b/torchci/rockset/commons/__sql/recent_pr_workflows_query.sql
@@ -6,7 +6,7 @@ WITH recent_shas AS (
     p.number AS number
   FROM
     workflow_job j
-    JOIN commons.pull_request p ON j.head_sha = p.head.sha
+    JOIN commons.pull_request p ON j.head_sha = p.head.sha HINT(join_strategy = lookup)
   WHERE
     (
       (
@@ -17,7 +17,7 @@ WITH recent_shas AS (
       )
       OR : prNumber = p.number
     )
-    AND p.base.repo.full_name = : repo
+    AND p.base.repo.full_name =: repo
 )
 SELECT
   w.id AS workflowId,
@@ -37,9 +37,11 @@ SELECT
   j.torchci_classification.context AS failure_context,
   j._event_time AS time
 FROM
-  recent_shas
-  JOIN commons.workflow_job j ON j.head_sha = recent_shas.sha
-  JOIN commons.workflow_run w ON w.id = j.run_id
+  commons.workflow_run w
+  JOIN (
+    commons.workflow_job j
+    JOIN recent_shas ON j.head_sha = recent_shas.sha HINT(join_strategy = lookup)
+  ) ON w.id = j.run_id HINT(join_strategy = lookup)
 UNION
 SELECT
   null AS workflowId,
@@ -59,7 +61,7 @@ SELECT
   null AS failure_context,
   w._event_time as time
 FROM
-  recent_shas
-  JOIN commons.workflow_run w ON w.head_sha = recent_shas.sha
+  commons.workflow_run w
+  JOIN recent_shas ON w.head_sha = recent_shas.sha HINT(join_strategy = lookup)
 ORDER BY
   time DESC