--- a/torchci/rockset/prodVersions.json
+++ b/torchci/rockset/prodVersions.json
@@ -23,7 +23,7 @@
     "issue_query": "e4d338de89980044",
     "failure_samples_query": "7940a636284d0752",
     "num_commits_master": "e4a864147cf3bf44",
-    "recent_pr_workflows_query": "2dd4ad8044ee6c65",
+    "recent_pr_workflows_query": "d42ec445f05da4d9",
     "reverted_prs_with_reason": "751f01cba16364f0",
     "unclassified": "1b31a2d8f4ab7230",
     "test_insights_overview": "42dbd5232f45fd53",