Some Dr. CI queries are failing due to too much resource utilization according to the new query_logs table

To improve speed + reduce memory utilization, I move tables in joins around to be larger table join smaller table and use the broadcast join strategy.  

Approximate improvements according to the query profiler:
All PRs (past 30 minutes):
max memory of a step: 37GB -> 500 MB
time: ~8s -> ~2s

Single PR (used 121717)
max memory of a step: 37GB -> 200 MB
time: ~6s -> ~1s