@huydhn 2024-03-13
This is nice.  Dr.CI also uses this, so this should be very useful optimization.  I feel that we should have jot down about how important a SQL query is like in the band of 1 (experiment) to 5 (critical) on the SQL comment, so that we pay more attention to their performance.  This one is probably a 4 or 5 IMO